+++ src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/domain/chat/adapter/in/ChatConsumeService.java
@@ -11,7 +11,7 @@ import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.chat.adapter.in
 import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.chat.adapter.in.mapping.ChatDtoMapping;
 import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.chat.entity.Chat;
 import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.chat.repository.ChatRepository;
-import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.chat.service.ChatRelayService;
+import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.chat.service.ChatDeliveryService;
 import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.chatparticipation.entity.ChatParticipation;
 import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.chatparticipation.repository.ChatParticipationRepository;
 import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.chatroom.entity.Chatroom;
@@ -34,7 +34,7 @@ public class ChatConsumeService {
 
     private static final Duration IDEMPOTENCY_TTL = Duration.ofMinutes(3);
 
-    private final ChatRelayService chatRelayService;
+    private final ChatDeliveryService chatDeliveryService;
     private final IdempotencyService idempotencyService;
     private final ChatRepository chatRepository;
     private final UserRepository userRepository;
@@ -63,7 +63,7 @@ public class ChatConsumeService {
 
             chatRepository.save(new Chat(sender, receiver, chatroom, chatDto.content()));
 
-            chatRelayService.relayOrNotify(receiver, ChatDtoMapping.consumeToRelay(chatDto));
+            chatDeliveryService.relayOrNotify(receiver, ChatDtoMapping.consumeToRelay(chatDto));
 
         } catch (Exception e) {
             if (idempotencyService.hasKey(idempotentKey)) {
