+++ src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/domain/chat/service/ChatRelayService.java
@@ -1,45 +1,24 @@
 package kakaotech.bootcamp.respec.specranking.chatconsumer.domain.chat.service;
 
-import static kakaotech.bootcamp.respec.specranking.chatconsumer.global.common.type.NotificationTargetType.CHAT;
-
 import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.chat.dto.ChatRelayDto;
-import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.notification.entity.Notification;
-import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.notification.repository.NotificationRepository;
-import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.user.entity.User;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.reactive.function.client.WebClient;
 
 @Service
-@Transactional
 @RequiredArgsConstructor
 public class ChatRelayService {
-
-    private final RedisTemplate<String, Object> redisTemplate;
     private final WebClient webClient;
-    private final NotificationRepository notificationRepository;
 
-    private static final String REDIS_USER_KEY_PREFIX = "chat:user:";
-    private static final String CHAT_RELAY_API_PATH = "/api/chat/relay";
     private static final String SCHEME = "http://";
+    private static final String CHAT_RELAY_API_PATH = "/api/chat/relay";
 
-    public void relayOrNotify(User receiver, ChatRelayDto dto) {
-        Object serverIpObj = redisTemplate.opsForValue().get(REDIS_USER_KEY_PREFIX + receiver.getId());
-
-        if (serverIpObj instanceof String serverIp && !serverIp.isEmpty()) {
-            webClient.post()
-                    .uri(SCHEME + serverIp + CHAT_RELAY_API_PATH)
-                    .bodyValue(dto)
-                    .retrieve()
-                    .bodyToMono(String.class)
-                    .subscribe();
-        } else {
-            if (notificationRepository.existsByUserIdAndTargetName(receiver.getId(), CHAT)) {
-                return;
-            }
-            notificationRepository.save(new Notification(receiver, CHAT));
-        }
+    public void relay(String serverIp, ChatRelayDto dto) {
+        webClient.post()
+                .uri(SCHEME + serverIp + CHAT_RELAY_API_PATH)
+                .bodyValue(dto)
+                .retrieve()
+                .bodyToMono(String.class)
+                .subscribe();
     }
 }
