+++ src/main/java/kakaotech/bootcamp/respec/specranking/chatconsumer/domain/notification/service/NotificationService.java
@@ -1,0 +1,21 @@
+package kakaotech.bootcamp.respec.specranking.chatconsumer.domain.notification.service;
+
+import static kakaotech.bootcamp.respec.specranking.chatconsumer.global.common.type.NotificationTargetType.CHAT;
+
+import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.notification.entity.Notification;
+import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.notification.repository.NotificationRepository;
+import kakaotech.bootcamp.respec.specranking.chatconsumer.domain.user.entity.User;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class NotificationService {
+    private final NotificationRepository notificationRepository;
+
+    public void createChatNotificationIfNotExists(User receiver) {
+        if (!notificationRepository.existsByUserIdAndTargetName(receiver.getId(), CHAT)) {
+            notificationRepository.save(new Notification(receiver, CHAT));
+        }
+    }
+}
